admt: Updated CNV change method - Changed CNV page to uint8_t - Added registry read for rotation, angle, count, temp channels - Added disable state for lineedit - Changed error value for double to UINT32_MAX

+++ plugins/admt/src/widgets/registerblockwidget.cpp
@@ -27,7 +27,7 @@
 using namespace scopy;
 using namespace scopy::admt;
 
-RegisterBlockWidget::RegisterBlockWidget(QString header, QString description, uint32_t address, uint32_t cnvPage,
+RegisterBlockWidget::RegisterBlockWidget(QString header, QString description, uint32_t address, uint8_t cnvPage,
 					 RegisterBlockWidget::ACCESS_PERMISSION accessPermission, QWidget *parent)
 	: QWidget(parent)
 	, m_address(address)
